Create Tabs for SampleDetails sections
